refactor: 전체적인 패키지 구조 계층화하여 변경
