+++ src/main/java/roomescape/domain/reservation/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.domain.reservation;
 
 import java.util.List;
