+++ src/main/java/roomescape/infrastructure/reservation/MemoryReservationRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
-import roomescape.model.Reservation;
-import roomescape.model.ReservationRepository;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationRepository;
 
 @Repository
 public class MemoryReservationRepository implements ReservationRepository {
